add spring security dependency WebConfig.java -> WebSecurityConfig.java add SecurityFilterChain to implement cors and set up permissive rules of access
