+++ boot/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation project(':exploration-module')
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-actuator'
+    implementation 'org.springframework.boot:spring-boot-starter-security'
     implementation 'org.mybatis.spring.boot:mybatis-spring-boot-starter'
     implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui'
