+++ build.gradle
@@ -121,6 +121,7 @@ subprojects {
             implementation "org.springframework.boot:spring-boot-starter-web:${springBootVersion}"
             implementation "org.springframework.boot:spring-boot-starter-integration:${springBootVersion}"
             implementation "org.springframework.boot:spring-boot-starter-actuator:${springBootVersion}"
+            implementation "org.springframework.boot:spring-boot-starter-security:${springBootVersion}"
             implementation "org.springframework.integration:spring-integration-http:${springIntegrationHttpVersion}"
             implementation "org.springframework.boot:spring-boot-starter-validation:${springBootVersion}"
             implementation "org.springdoc:springdoc-openapi-starter-webmvc-ui:${springdocOpenapiVersion}"
